+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/ObjectIdentifier.kt
@@ -5,6 +5,7 @@ import at.asitplus.signum.indispensable.asn1.encoding.decodeAsn1VarBigInt
 import at.asitplus.signum.indispensable.asn1.encoding.toAsn1VarInt
 import at.asitplus.signum.indispensable.asn1.encoding.toBigInteger
 import com.ionspin.kotlin.bignum.integer.BigInteger
+import kotlinx.io.Buffer
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -78,12 +79,12 @@ class ObjectIdentifier @Throws(Asn1Exception::class) private constructor(
                     collected += BigInteger.fromUInt(this.bytes[index].toUInt())
                     index++
                 } else {
-                    val currentNode = mutableListOf<Byte>()
+                    val currentNode = Buffer() //todo: calculate only index and then operate on a byte view (unsafe wrapped)
                     while (this.bytes[index] < 0) {
-                        currentNode += this.bytes[index] //+= parsed
+                        currentNode.writeByte(this.bytes[index]) //+= parsed
                         index++
                     }
-                    currentNode += this.bytes[index]
+                    currentNode.writeByte(this.bytes[index])
                     index++
                     collected += currentNode.decodeAsn1VarBigInt().first
                 }